requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0
pymongo==4.6.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# SSH & Remote Management
paramiko==3.4.0
asyncssh==2.14.1

# Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings[python]==0.24.0

# API Documentation
python-multipart==0.0.6
itsdangerous==2.1.2
